requirements.txt
h5py==3.8.0
louvain==0.8.2
MarkupSafe==2.1.5
matplotlib==3.5.3
matplotlib-inline==0.1.6
munkres==1.1.4
networkx==2.6.3
numpy
opencv-python==4.8.1.78
pandas==1.3.5
Pillow==9.3.0
PyYAML
rpy2==3.5.14
scanpy==1.9.3
scikit-learn
scikit-misc==0.1.4
scipy
seaborn==0.12.2
timm==0.9.10
torch==1.13.0+cu117
torch-geometric
torchaudio==0.13.0+cu117
torchvision==0.14.0+cu117
umap-learn==0.5.4